improve BlockRecordSource and remove override
